Move choice options enums from InnerSchedulers to InnerScheduler.Props
