+++ src/components/Scheduler/InnerSchedulers/InnerScheduler.Props.ts
@@ -1,7 +1,16 @@
-import {Schedule } from '../Scheduler.Props';
+import { Schedule } from '../Scheduler.Props';
 
 export interface IInnerSchedulerProps {
     schedule: Schedule;
     onScheduleChanged: (schedule: Schedule) => void;
 }
 
+export enum DailySheduleOptionEnum {
+    EveryAmountOfDays,
+    EveryWeekday
+}
+
+export enum MonthlySheduleOptionEnum {
+    OnDayOfMonth = '1',
+    OnDayOfWeek = '2'
+}
